requirements.txt
Flask
cloudinary
requests
python-dotenv